+++ NEWS.md
@@ -2,7 +2,7 @@
 
 ### Enhancements
  * Analysis functions (`cfun/afun`) have more information about current column split; `.spl_context` has access to `cur_col_id`, `cur_col_expr`, `cur_col_split`, and `cur_col_split_val`.
- * Analysis functions (`cfun/afun`) can use new parameters to extend analysis calculations to `alt_count_df`, i.e. `.alt_counts_df`, `.all_col_exprs`, and `.all_col_counts`.
+ * Analysis functions (`cfun/afun`) can use new parameters to extend analysis calculations. `.alt_counts_df` gives access to `alt_counts_df` across columns, while `.all_col_exprs` and `.all_col_counts` contains global information about all columns.
  * Added check for `.alt_counts_df` that prevents its calculation if not present in analysis functions.
 
 ### Bug Fixes